requirements.txt
pandas
duckdb
pyarrow
pyyaml
